Add snippet generator help for 'clean when status is'
Places the help inside the snippet generator page but does not place the
help inside the Pipeline steps reference.

The user interface for the checkboxes makes sense in the snippet
generator.  No need to add help for each checkbox in the snippet
generator.  Better to use a single help section to describe all the
checkboxes.